Constants Rearranged
Note: This is part 1 of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. As such, the code may not compile until the last commit in this series is made.

The constant colors in RambleyPainter have been rearranged to make more sense, along with some of the constants related to calculating the ears and the flags. Additionally, the second color in the background gradient has been made constant, and a teeth outline color has been specified. The internal render dimensions have been changed from integers to doubles, and the size of RambleyIcon has been updated accordingly. The colors, internal render dimensions, and some of the flags have been documented. The flags for ignoring aspect ratio and evil Rambley have been bit shifted to the left to make way for a flag for painting Rambley's border and shadow (name not yet set in stone). Documentation has been added to the flags field, the pixelGrid field, the listenerList field, and the iris and pupil fields. References to the Snake class from Snake4J, from which I initially copied the flag methods from, have been removed. The comment at the top outlining the initial idea has been moved to above the class documentation.