Dot + Grid Constants now Doubles, Rambley coord is now constant
Note: This is part 2 of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. As such, the code may not compile until the last commit in this series is made.

The background dot constants have changed from integers to doubles, and the BACKGROUND_DOT_POINTS constant has been removed since it is no longer needed. The background dot offsets are now calculated as doubles, and the method used to paint the background dots now use a Path2D to create the dots instead of a polygon, and the drawing routines for the background dots has been reworked to use double precision. The BACKGROUND_DOT_HALF_SIZE constant is now private. Incomplete documentation has been added to the getBackgroundGradient method. The PIXEL_GRID_SPACING constant is now a double, and the getPixelGrid method can now be given an optional Path2D object to reuse (a version that uses the pixelGrid Path2D object has also been made, but that one is private). This is mainly to allow the method to be overridden without having access to the pixelGrid field. The coordinates for the top-left corner at which Rambley will be rendered are now constants instead of a local variable