Iris and Pupil sizes are constant
Note: This is part 3 of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. As such, the code may not compile until the last commit in this series is made.

The size of the iris and pupil are now constant. Partial documentation has been added to the getRambleyStroke method along with the methods relating to Rambley's eyes. A version of paintRambleyEye has been added that takes in both the center of the iris and pupil, along with the ellipses for the iris and pupil.