Math And Debug Moved to the bottom of the file
Note: This is part 4 of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. As such, the code WILL not compile until the last commit in this series is made.

The math related functions have been moved to the bottom of the file and made protected and static. Documentation has been added to the getRadius and getCircleIntersections methods, along with partial documentation for the other math functions. The methods related to quadratic Bezier curves have been renamed. Getting the control point for a quadratic Bezier curve has been corrected to use the correct approach as laid out in this paper: https://microbians.com/math/Gabriel_Suchowolski_Quadratic_bezier_through_three_points_and_the_-equivalent_quadratic_bezier_(theorem)-.pdf Methods have been added to get the control points for a cubic Bezier curve, and for a spline formed by cubic Bezier curves. Methods has been added to brute force finding the intersection point of two lines, with one taking in a resolution, and the other using a default resolution value.

The debug code has also been moved to the bottom of the file under the math functions. The debug code is temporary and should not be relied on.