Rambley Ear Equations Renamed
Note: This is part 5 of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. As such, the code WILL not compile until the last commit in this series is made.

The functions for calculating the curves for the upper, lower, and tip portions of the ears have been renamed. The functions for converting the points to and from the coordinate space used by painter and the coordinate space used by the ear curve functions have also been renamed. The y offset for the tip is now a constant instead of retrieved from getTestDouble6. I've added methods for getting the y-coordinates for the ear curve functions as objects, and methods for getting the intersection between the upper curve and the tip and the intersection between the lower curve and the tip. A portion of the debug code was accidentally left out of the previous commit.