Rambley's Ear is completed
Note: This is the last part of a series of commits that will be getting made since a lot of changes have been made. These changes have been broken up to make the commits more concise. The code should compile after this commit.

The new code to generate the shape of Rambley's ear has been added, with the remnants of the old ear code being removed. RAMBLEY_EAR_Y_OFFSET_2 has been removed and incorporated into getting the y coordinate for the right ear. The method for getting the inner portion of Rambley's ear has been made protected instead of private. The scale for the inner ear has become a constant. Finally, the rendering hint for how to normalize strokes has been set to pure (i.e. the geometry should be left unmodified).