Added size control to test program
Added components and code to control the size at which RambleyIcon (and thus RambleyPainter) is rendered by the test program, along with rearranging some of the GUI components on the test program. I've also removed the "protected" modifiers from the debug methods of RambleyPainter (keep in mind, these methods will be removed when RambleyPainter is finished, and should not be used by anything other than the test program).