Added a setFlags method to RambleyPainter
A method to set all the flags at once with a single integer has been added to RambleyPainter. One was intended, but I forgot to add it. The existing methods for setting flags (setFlag and toggleFlag) have been updated to call the new setFlags method and the fireFlagChange method has been removed since it is no longer needed.