Rearranged and added documentation to fields
I've rearranged the fields in RambleyPainter to fall under 4 categories. These categories are Settings and Listeners, for fields that store settings, listeners, and other related stuff, Commonly Used Objects, for fields that don't change once initialized and that are used each time Rambley is painted, Dedicated Scratch Shape Objects, for scratch shape objects used to render specific things and only those things, and Generic Scratch Shape Objects, for scratch shape objects that are effectively free to use whenever for whatever purpose and are mostly here to reduce the amount of shape objects that are only used once when painting Rambley.

I've also added documentation for the DoubleUnaryOperators used for the ear calculations and the BasicStroke objects. The eye stroke has been renamed to the detail stroke, as it is used for more than just the eyes.