Added (currently unused) eye location fields
I've added fields to RambleyPainter to control the location of Rambley's eyes (more accurately the location of the iris and pupil of each eye, though the pupil's location is locked to the center of the iris). Each eye is to be controlled by two double values for the x and y components of the location, with 0.5 being the center of the eyes. Each eye will be able to be controlled separately, though methods to set them both at the same time will be provided. As of right now, these values do nothing and cannot be altered externally.