Setting changes can be chained + Documentation
I've changed it so that the methods used to set the flags now return the RambleyPainter that they were invoked upon, allowing changes to the settings to be chained together. This does mean that the setFlag and toggleFlag methods have now lost the ability to return whether a change was actually made to the settings, but this was mainly a holdover from the methods from the Snake4J Snake class from which these methods were copied from, as the Snake class did not make those methods public facing and used the returned values to indicate whether a property change event should be fired. I've also reworked the documentation for the flags (excluding the debug ones since those will be removed at a later time), along with adding documentation for the inner ear scale constant. I've also added documentation to the isBackgroundPainted method. Also, mentions of the background dots in the documentation have been changed to specify that they are polka dots, since I just remembered that's what the pattern is called. The pixel grid effect is now painted by default. I am currently considering whether to consider the flags as properties.