Rambley's shadow is painted by default
Rambley's border and shadow are now to be painted by default. This currently does nothing as the functionality to render the border and shadow aren't implemented yet, but when they are, they will be rendered by default. Additionally, I've corrected a spelling mistake, as antialiasing starts with an a, not a g.