Paint code creates copy Graphics2Ds
The code for the paintBackground, paintBackgroundDots, paintRambley, paintRambleyEye, and paintPixelGrid methods now create their own copies of the Graphics2D objects passed to them, negating the need for paint (paintBackground in the case of paintBackgroundDots and paintRambley in the case of paintRambleyEye) to provide copies of their Graphics2D objects to avoid unintended changes from occurring. Additionally, some documentation has been added to the getPixelGrid, configureGraphics, paintBackground, paintBackgroundDots, and paintPixelGrid methods along with the methods for getting the strokes used for drawing Rambley. The private version of getPixelGrid has been removed, since its functionality (which was calling the protected version of getPixelGrid with the pixelGrid field) has been incorporated into the paintPixelGrid method. The paint method now checks if the given Graphics2D object is null, and throws a NullPointerException if it is.