Rambley's head is centered horizontally now
Rambley's head is now centered horizontally, as the calculations for creating the head shape now use the center of the image and expand horizontally from there. It is not vertically centered, only horizontally centered. This also makes it so that the horizontal flip transform is no longer offset by 2 to account for Rambley being 1 pixel off center horizontally. The RAMBLEY_X_OFFSET has also been updated to be 28 instead of 27, though it currently is not used and may be removed later. The methods for getting the DoubleUnaryOperator versions of the Rambley ear equations are now documented and protected. The debug settings now no longer use the flags, instead using dedicated fields for whether lines are shown and whether we're doing A/B testing.