Added a Commonly Used Horizontal Flip Transform
I've made it so RambleyPainter now creates and stores an AffineTransform to flip things horizontally, instead of creating one each time paintRambley is called. The horizontal flip transform can be accessed using the getRambleyHorizontalFlipTransform method, though since it is most commonly used to flip areas, a method to flip areas using the transform has been added, that being the createHorizontallyFlippedArea method. Every instance of the local horizFlip has been replaced with either the getRambleyHorizontalFlipTransform method or a call to createHorizontallyFlippedArea on the area being transformed.