Rambley's head shape is now a function
I've moved the code for generating the shape of Rambley's head without ears to it's own function (currently a private function, though that may change later). Additionally, after some analysis I found that the Path2D head1a is only ever referenced once after being used to create the head shape, and that is to get its minimum y coordinate. As such, I have swapped it out with the scratch path object and just retrieved its minimum y coordinate after I created the head shape. The ears also now use the head bounds' minimum y instead of the head2 Ellipse2D's minimum y, as they are the same value and head2 may become a scratch Ellipse2D later down the line. Due to the changes in how the code works, some of the debug code for drawing the lines that make up Rambley have been moved to before the ears are created. The angle for the cheeks and the height of the triangle that forms the mask for the cheeks have become constants.