Rambley Ear Update
I've added documentation to the RAMBLEY_EAR_Y_OFFSET, RAMBLEY_EAR_TIP_Y_OFFSET, and RAMBLEY_EAR_TIP_ROUNDING constants. I've also added comments to the getRambleyEar method. Additionally the addQuadBezierCurve method is now protected static and has joined the other mathematical methods at the bottom of the file. Finally, the transition between the tip of Rambley's ear and the lower curve of Rambley's ear now has a smooth transition, similar to the transition between the upper curve of Rambley's ear and the tip of Rambley's ear. The difference is practically imperceptible, but in situations where it is, it should make it look a bit better.