Documentation and Reorganization
I've added some documentation to the methods and constants related to calculating the curves for Rambley's ears, along with some documentation and comments for the math methods. The private getSplineFirstCtrlPoints method has been incorporated into the getCubicBezierSplineControlPoints method, which was the only method that called it. The getLineIntersection method now waits until the end to offset the point by the desired x and y offsets, instead of using the offset values throughout, and uses scratch Line2D objects to reduce the amount of Line2D objects it creates. The addQuadBezierCurve now returns the control point that it calculates. The methods relating to calculating and creating Rambley's ears have been moved above the methods for Rambley's eyes (just below the method for creating Rambley's head shape).