Rambley's facial markings are now a function.
I've moved the code for generating the shape of Rambley's facial markings to it's own function (currently a private function, though that may change later). Additionally, after some analysis I found that the head3 Ellipse2D and head5 Rectangle2D are really only referenced within the code generating the facial markings (excluding the debug code since that doesn't count), and as such they are now replaced with the ellipse and rect scratch objects. The head4 Ellipse2D has been renamed to head3, and the head6 Ellipse2D has been renamed to head4. I've also made the notch ellipse (head4) to be based off the location of the main ellipse forming the markings instead of the triangle that forms the cheeks.