Consolidated Rambley Paint code
I've consolidated the code used to draw the shapes that makes up Rambley (when not showing lines for debug purposes) to all be in one location. This makes it easy to find the code responsible for drawing Rambley. I've also documented this code. The paintRambleyEye(Graphics2D, Shape, double, double) method has been removed since it is no longer needed (paintRambley initializes the iris and pupil Ellipse2D objects, and so I just made it so that the paintRambley method provides the two Ellipse2D objects to the version of the function that takes in two Ellipse2D objects for this exact purpose).