Ellipse method for getting points for a given x
I've added a method for getting the points on an ellipse for a given x-coordinate. This method is now used by the line versions of the  getCircleIntersections method when given a vertical line. Some documentation have been added to the getCircleIntersections methods, and the point based version of the line version of getCircleIntersections now returns a boolean indicating whether the line intersects the ellipse.