Rambley's Eye markings are now a function
I've made it so that Rambley's eye markings are generated by functions. It turns out that Rambley's eye markings didn't need eye4 and eye5 (these ARE used when creating Rambley's eyes). I also realized that I wasn't using the intersections between the snout (head5) and the top line of head1. Additionally, since the eye3 Path2D object is not used after this, the path scratch Path2D object is used instead. The point at which the path that forms the part of the eye markings not covered by the ellipse is now no longer roughly on the ellipse, and is now a point on the ellipse (it's still calculated by getting the rough points of intersection between the eyebrow ellipse and the eye markings ellipse, it's just the y-coordinate gets recalculated to put it onto the ellipse to ensure it falls on the eye markings ellipse).