Ellipse method for getting points for a given y
I've added a method for getting the points on an ellipse for a given y-coordinate (a companion method for getEllipseY). This method, after checking easy cases such as a y-coordinate out of bounds or a y-coordinate at the top, center, or bottom of the ellipse, invokes the line version of the getCircleIntersections method with a horizontal line at the given y-coordinate. I've also renamed the third point (point3) given to the createRambleyEyeMarkings to pointC since it gets the control point.

Additionally, I've reset all the test doubles to 1.0. This is just debug code, so it doesn't matter.