Rambley's Eye shape are now a function
I've made it so that Rambley's eye shape are generated by functions. It was found that the eye4 Ellipse2D object's maximum y-coordinate was the same as eye1's (the eyebrow Ellipse2D object), and so it was changed so that it was the same maximum y-coordinate in the method. The eye4 and eye5 Ellipse2D objects aren't referenced after this, so they could become scratch objects. Additionally, eye5 was only ever used as a reference for the creation of the eye, and did not need to actually be an Ellipse2D object, it just needed to be a RectangularShape. As such, they have been replaced with the ellipse Ellipse2D and rect Rectangle2D scratch objects, respectively. The eye6 Path2D object was only ever referenced during the construction of the eye, so it was swapped out with the path Path2D scratch object. The reference to the head4 Ellipse2D object was found to be unnecessary, since it unnecessarily bound one of the curves for the eyes to the ellipse used to produce the mask-like markings. As such, it was replaced with the center of the head shifted 14 pixels to the left. The instance of the getCircleIntersections method in the eye creation was replaced with a call to getEllipseX, since it was just a horizontal line being used to intersect the ellipse.