Fixed Mouth function
There was a mistake in createRambleyMouthCurve where the x-coordinate was used instead of the y-coordinate when attempting to get the points where a horizontal line intersects with the snout ellipse in order to get the points for the curve of the mouth. This was corrected on my end but I forgot to save the file before making a commit.