Reformatting And Simplifying RambleyPainter
I've reformatted some of the code for painting Rambley, renaming some of the shape parameters. I've reworked the code for generating the nose, removing the need for the head1 Rectangle2D object, allowing the rect Rectangle2D scratch object to be used to generate the shape of the head. Additionally, nose1 is now replaced with rect, since it's not used much after rect is reused as a mask. It was used for the point for the mouth, but the bounds of the mouth is now used instead. This is in preparation for the nose generation code to be moved into a method.