Simplifying Rambley Nose code
I've simplified the code for Rambley's nose further, finding out that the nose1 Ellipse2D object wasn't really needed, as other things could be used instead. With nose1 removed, rect did not need to be used as a mask for nose1, and thus it continued to contain the bounds for the nose. As such, noseBounds was no longer needed and removed. The nose5 Path2D object was removed and replaced with the path Path2D scratch object after it is reset. The nose2 Ellipse2D object was removed and replaced with the ellipse Ellipse2D scratch object. Additionally the path used to form the bottom part of the nose is now flipped and added to the path instead of its area being flipped and added to the nose area. A method for creating AffineTransforms that flip horizontally and translate a given x offset has been added. Another method was added to horizontally mirror a path and adds the mirror back to the path. The method used to generate Rambley's mouth now uses the latter method to create the left side of the mouth.