Background polka dots can be circular now
The background polka dots can now be circles instead of diamonds now. Methods for changing whether the background polka dots are circles or diamonds have been added, as well as the methods for getting the background polka dots now take in an Ellipse2D object in addition to the parameters they already were taking in, and now return a shape object, which will either be the Path2D or the Ellipse2D object, depending on the settings. The paintBackgroundDots method now initializes the path Path2D scratch object and the ellipse1 Ellipse2D scratch object, and now passes the result of calling getBackgroundDot directly to Graphics2D.fill(Shape).