Mouth open parameters are added to RambleyPainter
Fields and methods for controlling how far and wide Rambley's mouth is open have been added, along with flags controlling the side on which Rambley's fang appears on and whether Rambley is showing his teeth. These currently do nothing, but it's a start to having an open mouth for Rambley. Additionally, the eye controls have been added to the paramString printout.