Update RambleyPainter.java
I've added a version of the createRambleyMouthCurve method that takes in 5 Point2D objects instead of 4, with the 1st one for that method to be effectively a Point2D with the x and y coordinate for the center spike of the mouth (bottom center of the nose). The paintRambley method now calls this new version, having shifted the Point2D objects it already used to the next one over, and point1 is now used to store the bottom center point of the nose. This is in preparation for code to allow Rambley to open his mouth.