Update RambleyPainter.java
I've renamed the mouthTop Path2D object to mouthPath. I've also changed references to Rambley showing his teeth to say Rambley closing his jaw, since that is a better fit for the description, renaming the related constants and methods. The open mouth wideness and height require their parameters to be within 0 and 1, inclusive, and now throw an exception if that is not the case.