Tongue and fang can be controlled by the open mouth
The tongue and fang methods now take in the same mouthWidth and mouthHeight parameters that are passed to createRambleyOpenMouthShape to control the shape of the open mouth, allowing them to be controlled by these values too. Additionally the createRambleyFangShape takes in the mid-point of the mouth curve. The tongue's y-coordinates are now adjusted by the open mouth's height, making it so smaller open mouths now have the tongue positioned lower than larger open mouths. There are plans to possibly shift the fang inward in a similar way, but I'm still thinking about whether that is even necessary. Additionally some line debug code has been removed.