Geometry Math is in its own library
I've added a library, that being GeometryMath, to handle the geometric math used throughout RambleyPainter, and moved the relevant methods and constants to that library. The getQuadBezierT now returns a double instead of an array of doubles, since it only really ever returned a single double, or at least the smaller of the two was ever used. GeometryMath also introduces two new methods, getQuadBezierPointForX and getQuadBezierPointForY, which call getQuadBezierT to get a t value for their given x or y, respectively, and then call getQuadBezierPoint to get the point at that t value. Instances where getQuadBezierT was called only for that t value to be given to getQuadBezierPoint have been replaced with their respective getQuadBezierPointFor method.  I've also added notes for myself to complete more of the documentation and added blank documentation to methods that didn't have one before.