Documentation + Separate pixel grid offset methods
I've added documentation to the methods that get the offsets for the background polka dots and the pixel grid effect. I've also added separate methods for getting the x and y offsets for the pixel grid effect for the same reason why there are separate methods for the x and y offsets of the background polka dots, that being so that subclasses could change these values independently by altering the methods that return the offsets. The non-axis dependent getBackgroundDotOffset and getPixelGridOffset methods have been made private.