Background Diamonds Vs Circles Update
I've changed the version of the getBackgroundDot method that takes in a RectangularShape bounds, a Path2D, and an Ellipse2D to always return the diamond shape. This version no longer takes in an Ellipse2D and has been renamed to getBackgroundDotDiamond. The other remaining getBackgroundDot method is now responsible for checking if the background dots are to be diamonds or circles, but no longer takes in a Rectangle2D object as it uses the Ellipse2D to set the bounds for a diamond instead in order to simplify things.