Documentation update for RambleyPainter
Minor change first: I've changed BACKGROUND_GRADIENT_COLOR_2 from a public constant to a protected constant. The main part of this change is a lot of documentation has been added or changed. To finish the documentation for the DEFAULT_BACKGROUND_DOT_SPACING constant, I directed people reading the documentation to go to the getBackgroundDotSpacing method for more information on what the background dot spacing does. This way once I rework the documentation for that method, I don't have to go back and rework the documentation for that constant. I've added a description of the gradient to the getBackgroundGradient method. I've added documentation stating the default implementations of and related methods to the getBackgroundDotOffsetX, getBackgroundDotOffsetY, getBackgroundDotOffset, getPixelGridOffsetX, getPixelGridOffsetY, and getPixelGridOffset methods. I've added a description and related methods to the getPixelGrid and getRambleyStroke methods. I've refined the descriptions of the BasicStrokes returned by the getRambleyNormalStroke, getRambleyDetailStroke, and getRambleyOutlineStroke methods. The documentation for the configureGraphics method now lists the related methods. The paintBackground, paintBackgroundDots, and the two paintPixelGrid methods have had the documentation and related methods added. The mask version of paintPixelGrid now has the masking code un-commented out. An unfinised version of the documentation for paintRambley has been added. A prototype for the method for creating the line line for the teeth when Rambley's jaw is closed has been added to the comments.