AffineTransform reuse and Horizontal Mirroring
I've added the ability to reuse AffineTransforms when getting a horizontal flip transform by passing an AffineTransform to the getHorizontalFlipTransform method. Two scratch AffineTransform objects have been added, one for flipping shapes horizontally on a case by case basis, and one for scaling the ears to get the inner ears. The getRambleyInnerEar now reuses the latter AffineTransform. I've also added a method for getting the RectangularShape that represents the bounds of a shape, in order to make it so that when RectangularShape methods are required for shapes, I can avoid creating an unnecessary Rectangle2D object when I could cast the shape to use the methods (this method is getBoundsOfShape). The code for altering Rambley's fang depending on whether Rambley is evil or his fang is on the left is now incorporated into the method to create the shape for the fang, with Evil Rambley's left fang being a mirror of the right fang path that is added to the fang path. A method has been added for getting a horizontal flip transform that translates over the maximum x-coordinate of a shape and a given x-coordinate has been provided, so as to get a transform that moves stuff back into the image. Another method, this one for creating an AffineTransform for horizontally mirroring a shape has also been added, since the previous way of creating a mirroring transform does not actually create a mirroring transform. The mirrorPathHorizontally method now uses the previously mentioned horizontal mirror transform to mirror paths correctly, and the new createHorizontallyMirroredArea uses the horizontal mirror transform to flip areas. Both of the previous two methods use the new scratch horizTx AffineTransform.