Flipped areas -> Mirrored areas
I've changed all instances of the use of createHorizontallyFlippedArea to create a flipped area to instead use createHorizontallyMirroredArea with an appropriate mirroring vertical line to flip the area. For most of these, it is the center of the head, though in the two cases that it is not, the x-coordinate for the mirroring line is a point that at one point was set using the center of the head bounds. This removes the assumption that the head is in the exact center of the image from the areas that were flipped. Additionally, the createHorizontallyFlippedArea and getRambleyHorizontalFlipTransform methods have been removed, along with the horizFlip AffineTransform object, since the first method is no longer used and worked on the assumption that we want to mirror over to the other side of the image, the second method was no longer used after the removal of the first method, and the horizFlip object was removed since the getRambleyHorizontalFlipTransform method was removed.