Start of Jaw closed implementation
I've begun the implementation of how it will draw Rambley with his jaw closed. While it is incomplete, as I need to make it so the line separating the two halves of the jaw is created and drawn, when Rambley's jaw is closed the inside of his mouth is no longer visible.