Rambley's head uses offsets again
I've made Rambley's earless head shape once again use the RAMBLEY_X_OFFSET value instead of the center of the image. Since the code no longer assumes that Rambley's head is centered in the image, the x-coordinate of Rambley's head is now irrelevant to the rest of the code. Additionally, the private createRambleyHeadArea method has been renamed to getRambleyEarlessHead, made protected, and takes in an x and y coordinate for the top-left corner of the earless head. This will allow his head to be repositionable in the future.