AffineTransform Objects now Dedicated scratch objects
The horizTx and inEarTx AffineTransform objects have been moved into the Dedicated scratch objects category from the Commonly used object category. This was previously a mistake, though it is a minor one, and only affects the clarification of stuff. I've also added a generic scratch AffineTransform object to be used whenever. Additionally, I've added descriptions to the field categories, and added disclaimers to the generic scratch objects stating that they can change at any time and may not be assumed to be in a known state before use.