Inner ear transform created by method
I've moved the code to create the AffineTransform  used to transform the outer ear shape into the inner ear shape to a separate method to create the transform. This is so that it might be reused or overridden by a subclass. The getRambleyInnerEar calls this new method to get the transform.