The inner ear transform is generic now
In a bit of foreshadowing, it turns out the transform for the inner ear will need to be used for more than just the inner ear. As such, the getRambleyInnerEarTransform method has been renamed to getCenteredScaleTransform and become a more generic method in the documentation. Additionally, the scale parameter has been removed from the getRambleyInnerEar method, and instead it just uses RAMBLEY_INNER_EAR_SCALE for the scale factor.