Invalid flags check and Rambley offset in method
I've added a maximum value for the flag and checks for whether the maximum flag would be out of range. Additionally the RAMBLEY_X_OFFSET and RAMBLEY_Y_OFFSET constants are now private, and are returned by the methods getRambleyX and getRambleyY, so as to allow them to be overridden by subclasses or possibly made public and variable in the future.