Documentation and Simplification
I've removed the version of createRambleyMouthCurve that took in an x and a y coordinate, leaving the version that takes a Point2D for the center point of the mouth curve. I've also updated the documentation for many of the methods, adding the related methods to the documentation of more methods. I've also finished the documentation for the paintRambleyBorderAndShadow method.