Major update to createRambleyFangShape
The getRambleyFangCurve method is no longer called by the createRambleyFangShape method, and must instead be called before calling the createRambleyFangShape method, with the createRambleyFangShape taking in the points calculated by getRambleyFangCurve. This is in preparation for the line that appears when Rambley's jaw is closed, since it will also need these points. Additionally, createRambleyFangShape now takes in the area of the open mouth, it is now responsible for intersecting the open mouth shape with the fang area, so as to remove all but what lies within Rambley's mouth from the fangs. The documentation for the createRambleyFangShape has been reworked to state the changes, and to state that it doesn't just produce the right fang. Additionally, the documentation for the getRambleyFangCurve method has been mostly completed, and the documentation for methods related to Rambley's mouth now reference the getRambleyFangCurve method along with an upcoming createRambleyClosedTeethLine method that is not implemented yet. Finally, the code in paintRambley that is responsible for creating and drawing the fang(s) has been changed so that it calculates the curve for Rambley's right fang regardless of whether Rambley's jaw is closed, but it will only generate and draw the shape for Rambley's fang(s) if Rambley's jaw is open.