Update RambleyPainter.java
I've made the RAMBLEY_EAR_HEIGHT constant protected instead of private, and added a note to its documentation stating that it is not the exact height of the ears. I've also added a reference to the RAMBLEY_EAR_HEIGHT and RAMBLEY_EAR_TIP_ROUNDING constants to the getRambleyEar method's documentation, and references to the drop shadow constants to the paintRambleyBorderAndShadow method's documentation.