getRambleyFangCurve takes in 1 reference point now
The getRambleyFangCurve takes in a single x and y coordinate on which to base the location of the fang, this being the location of the top center of the visible portion. Constants have been added for the width and height of the fang. As a result of these changes, createRambleyFangShape now once again calls getRambleyFangCurve to calculate the curve for the fang instead of receiving the points for the curve from paintRambley, and the call to getRambleyFangCurve in paintRambley has been removed. This change will make it easier to have a fang drawn on the closed jaw, as that fang may need to be in a different position and may need to calculate both sides of the fang curve instead of mirroring them. This also means that the size and curve of the fang is no longer dependent on mouth curve.