Update GeometryMath.java
I've added versions of the methods for dealing with quadratic bezier curves that take in points as individual x and y values, along with versions that take in a QuadCurve2D object. The methods effected by this change are the getQuadBezierPoint, getQuadBezierPointForX, and getQuadBezierPointForY methods. I've also added a method called getQuadBezierCurve which is effectively a version of getQuadBezierControlPoint that takes in and returns a QuadCurve2D object instead of a Point2D object. There is also now a getCubicBezierCurve method that, like getQuadBezierCurve for getQuadBezierControlPoint, is a CubicCurve2D version of the getCubicBezierControlPoints method. There is also a getCubicBezierSplineCurves method, which is a version of getCubicBezierSplineControlPoints that returns a list of CubicCurve2D objects. Additionally, the getCubicBezierSplineControlPoints method now returns the number of splines.