Rambley teeth line added, not finished
I've added the createRambleyClosedTeethLine method to create the line to draw on Rambley's mouth when his jaw is closed. The method is still in development, and what it generates is subject to change. The parameters are also subject to change, as currently it takes in more than it will likely need when finished.