Mouth edge lines now curve
I've made it so that the lines at the edges of Rambley's mouth are now curved. I've also made it so that the path used to produce the open mouth recreates the mouth curve at the top as much as it needs to instead of adding the existing mouth curve to the path. Additionally, I've added a method to get whether Rambley's mouth is open.