Mouth curve is higher when mouth is open
The lowest point of the top mouth curve is now higher, meaning that the curve is thinner vertically, when Rambley's mouth is open. If the mouth is closed, then the lowest point is 9 pixels below the first point. If the mouth is open to 25% of its height or more, then the lowest point will be 6 pixels below the first point. If the mouth is open somewhere between 0% and 25% of its height, then it will be proportional to the height of the mouth. As a result, I've made the visible portion of Rambley's fang to be 1.5 pixels taller, and it's now 1 pixel over to the left when the jaw is open.