Rambley's closed teeth line is finished
The createRambleyClosedTeethLine method has been finished (excluding documentation). The method now produces a more game-accurate line. paintRambley now uses a copy of the graphics context to render the inside of Rambley's mouth (or the teeth line if Rambley's mouth is closed). The debug code for Rambley's mouth has been removed. The point9 Point2D object was deemed unnecessary and removed, and the quadCurve2 QuadCurve2D object may be removed in the future.