Update RambleyPainter.java
The Rambley x offset is now calculated in a way that should center the head. The mouthCurve3 QuadCurve2D object was removed, instead using the quadCurve2 scratch object. I've added an Area that will get the full outline of Rambley, to be used to draw the border and shadow. I've made preparations for Rambley's scarf to be added. I've also made it so that Rambley's head and face are rendered to a copy of the graphics context. I've added comments indicating where future parts will go in paintRambley.