More accurate Rambley snout
I've made the Rambley snout area more accurate when the eyes aren't shown. While this is isn't visible because it's covered by the eyes (for now), it is more accurate to the design. Additionally, the bottom-left curve for Rambley's eye markings is now 12 pixels lower than the eye markings ellipse instead of 14.